perspective projection and lookat matrices.
Former-commit-id: 21dede8bb8d6c3cc7de091ff733752d0dd48d53a

+++ src/rffalcon/Block.cpp
@@ -46,7 +46,7 @@ void Block::render()
 
 void Block::getMCBoundingBox(double* xyzBounds) const
 {
-	int buffer = 0.25;
+	double buffer = 0.25;
 	xyzBounds[0] = _xmin - buffer;
 	xyzBounds[1] = _xmax + buffer;
 	xyzBounds[2] = _ymin - buffer;
